src/requirements.txt
aniso8601==8.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-Script==2.0.6
idna==2.10
influxdb==5.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==0.6.1
numpy==1.18.5
opencv-python==4.2.0.34
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==3.13
pyzbar==0.1.8
requests==2.24.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
